Changes in version 1 (Unfortunately will break things for v0 compiler users):
Files are now wrapped in individual functions at compile time and their @import-ed files are injected as parameters. At the end of the wrapping function, the file name is returned as a variable (which is assumed to identify the defined object variable in the file.) The wrapping function is then called and assigned to the file's namespace.

This eliminates the need to specify an object's namespace when defining it in a file, allowing the file to be used as would a normal js script.

Import statements now also allow for using the imported objects simple (non-namespaced) name in the context of the file.